added real picuters

+++ Online_Shop/Models/BeanBag.cs
@@ -10,7 +10,7 @@ namespace Online_Shop.Models
         public int id { get; set; }
         public string name { get; set; }
         public string description { get; set; }
-        public int price { get; set; }
+        public decimal price { get; set; }
         public string color { get; set; }
         public string measure { get; set; }
         public string picture { get; set; }
